+++ vet_calc_app/lib/ui/dosage_page.dart
@@ -16,83 +16,206 @@ class DosagePage extends StatelessWidget {
     return event.snapshot.value;
   }
 
-  //TODO: updated database to have dosages be in an object with
-  //      [{name (if available), low, high, unit}...]
-  //      iterate over list and calculate dosage for each dosage range for
-  //      dosage function
-  //      update concentration function to account for multiple dosages
+  //TODO: update concentration function to account for multiple dosages
 
   //calculates dosage
   //takes in weight, high dosage, low dosage, units, and if weight is in lb or kg
   //returns list with [dosage string, unit string, calculated low dosage, calculated high dosage]
-  List getDosage(double weight, double highDosage, double lowDosage,
-      String units, bool isLbs) {
+  List getDosage(double weight, var dosageList, bool isLbs) {
+    //counter to add or's in the string
+    var orCounter = 1;
+
     // convert to kg if lbs
     if (isLbs) {
       weight = weight * 0.453592;
     }
 
-    //if theres only one dosage
-    int kgIdx = units.indexOf("kg");
-    String beforeKgString = units.substring(0, kgIdx - 1);
-    String afterKgString = units.substring(kgIdx + 2);
+    String dosageString = "";
+    var dosageArray = [];
 
-    String unitSubstring = beforeKgString + afterKgString;
+    // loop through every dosage in the list
+    for (var object in dosageList) {
+      //check to see if dosage has a name for seperated dosages/concentrations
+      var name;
+      object['name'] == null ? name = null : name = object['name'];
 
-    // debugPrint("unitSubstring: $unitSubstring");
+      double highDosage = object['high'].toDouble();
+      double lowDosage = object['low'].toDouble();
 
-    //contains substring of just units and stuff after kg
-    if (highDosage == lowDosage) {
-      double dosage = weight * highDosage;
+      //extract the kg fro mthe units so it only has useful units
+      int kgIdx = object['units'].indexOf("kg");
+      String beforeKgString = object['units'].substring(0, kgIdx - 1);
+      String afterKgString = object['units'].substring(kgIdx + 2);
 
-      //rounding to 5 for now
-      return [
-        "${dosage.toStringAsFixed(3)} $unitSubstring",
-        unitSubstring,
-        dosage,
-        dosage
-      ];
-    } else {
-      double low = (weight * lowDosage);
-      double high = weight * highDosage;
+      String unitSubstring = beforeKgString + afterKgString;
 
-      return [
-        "${low.toStringAsFixed(3)} - ${high.toStringAsFixed(3)} $unitSubstring",
-        unitSubstring,
-        low,
-        high
-      ];
+      //calculate the dosage
+      if (highDosage == lowDosage) {
+        double dosage = weight * highDosage;
+
+        //rounding to 3 to fix double to string conversions
+        name == null
+            ? dosageString =
+                "$dosageString ${dosage.toStringAsFixed(3)} $unitSubstring\n"
+            : dosageString =
+                "$dosageString $name: ${dosage.toStringAsFixed(3)} $unitSubstring\n";
+        dosageArray.add({
+          'low': dosage,
+          'high': dosage,
+          'units': unitSubstring,
+          'name': name
+        });
+      } else {
+        double low = (weight * lowDosage);
+        double high = weight * highDosage;
+
+        //rounding to 3 to fix double to string conversions
+        name == null
+            ? dosageString =
+                "$dosageString ${low.toStringAsFixed(3)} - ${high.toStringAsFixed(3)} $unitSubstring\n"
+            : dosageString =
+                "$dosageString $name: ${low.toStringAsFixed(3)} - ${high.toStringAsFixed(3)} $unitSubstring\n";
+        dosageArray.add(
+            {'low': low, 'high': high, 'units': unitSubstring, 'name': name});
+      }
+
+      //add "or" if the dosages are not seperated from concentrations
+      if (orCounter != dosageList.length && name == null) {
+        dosageString = "$dosageString or\n";
+      }
+      orCounter++;
     }
+    return [dosageString, dosageArray];
   }
 
-  String getConcentration(
-      var concentration, var lowDosage, var highDosage, String dosageUnits) {
+  String getConcentration(var concentration, var dosageArray) {
     String concentrationString = "";
-    bool sameDosage = true;
-    if (lowDosage != highDosage) {
-      sameDosage = false;
-    }
+    //counter to add or's in the string
+    var orCounter = 1;
+
+    //iterate over concentrations
     for (var object in concentration) {
+      //get units out of concenration
       String concentrationUnits =
           object["units"].substring(0, object["units"].indexOf("/"));
       String concentrationAfterUnits = object["units"]
           .substring(object["units"].indexOf("/") + 1, object["units"].length);
 
-      String newDosageUnits = "";
+      //iterate over dosages
+      for (var dosage in dosageArray) {
+        var lowDosage = dosage['low'];
+        var highDosage = dosage['high'];
+        bool sameDosage = true;
 
+        if (lowDosage != highDosage) {
+          sameDosage = false;
+        }
+
+        //get units out of dosage
+        String newDosageUnits = "";
+        //see if dosage units are mg, µg or mEq
+        if (dosage['units'].contains("mg")) {
+          newDosageUnits = "mg";
+        } else if (dosage['units'].contains("µg")) {
+          newDosageUnits = "µg";
+        } else if (dosage['units'].contains("mEq")) {
+          newDosageUnits = "mEq";
+        }
+
+        double lowConcentration = object["low"].toDouble();
+        double highConcentration = object["high"].toDouble();
+        double lowConcentrationConverted = object["low"].toDouble() * 1000.0;
+        double highConcentrationConverted = object["high"].toDouble() * 1000.0;
+
+        //if the concentration is the same
+        if (lowConcentration == highConcentration) {
+          //if concentration units are mg while dosage units are µg
+          if (concentrationUnits == "mg" && newDosageUnits == "µg") {
+            if (sameDosage) {
+              concentrationString =
+                  "$concentrationString ${(lowDosage / lowConcentrationConverted).toStringAsFixed(5)} $concentrationAfterUnits\n";
+            } else {
+              concentrationString =
+                  "$concentrationString ${(lowDosage / lowConcentrationConverted).toStringAsFixed(5)} - ${(highDosage / lowConcentrationConverted).toStringAsFixed(5)} $concentrationAfterUnits\n";
+            }
+          } else {
+            if (sameDosage) {
+              concentrationString =
+                  "$concentrationString ${(lowDosage / lowConcentration).toStringAsFixed(5)} $concentrationAfterUnits\n";
+            } else {
+              concentrationString =
+                  "$concentrationString ${(lowDosage / lowConcentration).toStringAsFixed(5)} - ${(highDosage / lowConcentration).toStringAsFixed(5)} $concentrationAfterUnits\n";
+            }
+          }
+        } else {
+          if (concentrationUnits == "mg" && newDosageUnits == "µg") {
+            if (sameDosage) {
+              concentrationString =
+                  "$concentrationString ${(lowDosage / lowConcentrationConverted).toStringAsFixed(5)} - ${(lowDosage / highConcentrationConverted).toStringAsFixed(5)} $concentrationAfterUnits\n";
+            } else {
+              concentrationString =
+                  "$concentrationString ${(lowDosage / lowConcentrationConverted).toStringAsFixed(5)} - ${(lowDosage / highConcentrationConverted).toStringAsFixed(5)} to ${(highDosage / lowConcentrationConverted).toStringAsFixed(5)} - ${(highDosage / highConcentrationConverted).toStringAsFixed(5)} $concentrationAfterUnits\n";
+            }
+          } else {
+            if (sameDosage) {
+              concentrationString =
+                  "$concentrationString ${(lowDosage / lowConcentration).toStringAsFixed(5)} - ${(lowDosage / highConcentration).toStringAsFixed(5)} $concentrationAfterUnits\n";
+            } else {
+              concentrationString =
+                  "$concentrationString ${(lowDosage / lowConcentration).toStringAsFixed(5)} - ${(lowDosage / highConcentration).toStringAsFixed(5)} to ${(highDosage / lowConcentration).toStringAsFixed(5)} - ${(highConcentration).toStringAsFixed(5)} $concentrationAfterUnits\n";
+            }
+          }
+        }
+      }
+      if (orCounter != concentration.length) {
+        concentrationString = "$concentrationString or\n";
+      }
+      orCounter++;
+    }
+    return concentrationString;
+  }
+
+  String getConcentrationSeperated(var concentration, var dosageArray) {
+    String concentrationString = "";
+
+    for (var i = 0; i < concentration.length; i++) {
+      String name = "";
+      concentration[i]['name'] != null
+          ? name = "${concentration[i]['name']}:"
+          : name = "";
+
+      var lowDosage = dosageArray[i]['low'];
+      var highDosage = dosageArray[i]['high'];
+      bool sameDosage = true;
+
+      if (lowDosage != highDosage) {
+        sameDosage = false;
+      }
+
+      //get units out of concenration
+      String concentrationUnits = concentration[i]["units"]
+          .substring(0, concentration[i]["units"].indexOf("/"));
+      String concentrationAfterUnits = concentration[i]["units"].substring(
+          concentration[i]["units"].indexOf("/") + 1,
+          concentration[i]["units"].length);
+
+      //get units out of dosage
+      String newDosageUnits = "";
       //see if dosage units are mg, µg or mEq
-      if (dosageUnits.contains("mg")) {
+      if (dosageArray[i]['units'].contains("mg")) {
         newDosageUnits = "mg";
-      } else if (dosageUnits.contains("µg")) {
+      } else if (dosageArray[i]['units'].contains("µg")) {
         newDosageUnits = "µg";
-      } else if (dosageUnits.contains("mEq")) {
+      } else if (dosageArray[i]['units'].contains("mEq")) {
         newDosageUnits = "mEq";
       }
 
-      double lowConcentration = object["low"].toDouble();
-      double highConcentration = object["high"].toDouble();
-      double lowConcentrationConverted = object["low"].toDouble() * 1000.0;
-      double highConcentrationConverted = object["high"].toDouble() * 1000.0;
+      double lowConcentration = concentration[i]["low"].toDouble();
+      double highConcentration = concentration[i]["high"].toDouble();
+      double lowConcentrationConverted =
+          concentration[i]["low"].toDouble() * 1000.0;
+      double highConcentrationConverted =
+          concentration[i]["high"].toDouble() * 1000.0;
 
       //if the concentration is the same
       if (lowConcentration == highConcentration) {
@@ -100,36 +223,36 @@ class DosagePage extends StatelessWidget {
         if (concentrationUnits == "mg" && newDosageUnits == "µg") {
           if (sameDosage) {
             concentrationString =
-                "$concentrationString ${(lowDosage / lowConcentrationConverted).toStringAsFixed(5)} $concentrationAfterUnits\n";
+                "$concentrationString $name ${(lowDosage / lowConcentrationConverted).toStringAsFixed(5)} $concentrationAfterUnits\n";
           } else {
             concentrationString =
-                "$concentrationString ${(lowDosage / lowConcentrationConverted).toStringAsFixed(5)} - ${(highDosage / lowConcentrationConverted).toStringAsFixed(5)} $concentrationAfterUnits\n";
+                "$concentrationString $name ${(lowDosage / lowConcentrationConverted).toStringAsFixed(5)} - ${(highDosage / lowConcentrationConverted).toStringAsFixed(5)} $concentrationAfterUnits\n";
           }
         } else {
           if (sameDosage) {
             concentrationString =
-                "$concentrationString ${(lowDosage / lowConcentration).toStringAsFixed(5)} $concentrationAfterUnits\n";
+                "$concentrationString $name ${(lowDosage / lowConcentration).toStringAsFixed(5)} $concentrationAfterUnits\n";
           } else {
             concentrationString =
-                "$concentrationString ${(lowDosage / lowConcentration).toStringAsFixed(5)} - ${(highDosage / lowConcentration).toStringAsFixed(5)} $concentrationAfterUnits\n";
+                "$concentrationString $name ${(lowDosage / lowConcentration).toStringAsFixed(5)} - ${(highDosage / lowConcentration).toStringAsFixed(5)} $concentrationAfterUnits\n";
           }
         }
       } else {
         if (concentrationUnits == "mg" && newDosageUnits == "µg") {
           if (sameDosage) {
             concentrationString =
-                "$concentrationString ${(lowDosage / lowConcentrationConverted).toStringAsFixed(5)} - ${(lowDosage / highConcentrationConverted).toStringAsFixed(5)} $concentrationAfterUnits\n";
+                "$concentrationString $name ${(lowDosage / lowConcentrationConverted).toStringAsFixed(5)} - ${(lowDosage / highConcentrationConverted).toStringAsFixed(5)} $concentrationAfterUnits\n";
           } else {
             concentrationString =
-                "$concentrationString ${(lowDosage / lowConcentrationConverted).toStringAsFixed(5)} - ${(lowDosage / highConcentrationConverted).toStringAsFixed(5)} to ${(highDosage / lowConcentrationConverted).toStringAsFixed(5)} - ${(highDosage / highConcentrationConverted).toStringAsFixed(5)} $concentrationAfterUnits\n";
+                "$concentrationString $name ${(lowDosage / lowConcentrationConverted).toStringAsFixed(5)} - ${(lowDosage / highConcentrationConverted).toStringAsFixed(5)} to ${(highDosage / lowConcentrationConverted).toStringAsFixed(5)} - ${(highDosage / highConcentrationConverted).toStringAsFixed(5)} $concentrationAfterUnits\n";
           }
         } else {
           if (sameDosage) {
             concentrationString =
-                "$concentrationString ${(lowDosage / lowConcentration).toStringAsFixed(5)} - ${(lowDosage / highConcentration).toStringAsFixed(5)} $concentrationAfterUnits\n";
+                "$concentrationString $name ${(lowDosage / lowConcentration).toStringAsFixed(5)} - ${(lowDosage / highConcentration).toStringAsFixed(5)} $concentrationAfterUnits\n";
           } else {
             concentrationString =
-                "$concentrationString ${(lowDosage / lowConcentration).toStringAsFixed(5)} - ${(lowDosage / highConcentration).toStringAsFixed(5)} to ${(highDosage / lowConcentration).toStringAsFixed(5)} - ${(highConcentration).toStringAsFixed(5)} $concentrationAfterUnits\n";
+                "$concentrationString $name ${(lowDosage / lowConcentration).toStringAsFixed(5)} - ${(lowDosage / highConcentration).toStringAsFixed(5)} to ${(highDosage / lowConcentration).toStringAsFixed(5)} - ${(highConcentration).toStringAsFixed(5)} $concentrationAfterUnits\n";
           }
         }
       }
@@ -171,26 +294,33 @@ class DosagePage extends StatelessWidget {
             }
 
             // dosage
-            // returns dosage in [0] and units in [1]
-            var dosageList = getDosage(
-                weight.toDouble(),
-                mainDrug['Dosage_high'].toDouble(),
-                mainDrug['Dosage_low'].toDouble(),
-                mainDrug['Units'],
-                isLbs);
+            // see if its a string to just output that or go to function
 
-            var dosageDisplay = dosageList[0];
+            // getDosage returns dosage string in [0] and dosage array in [1]
+            // for concentration
+            String dosageDisplay = "";
+            List dosageList = [];
+            if (mainDrug['Dosage'] is String) {
+              dosageDisplay = mainDrug['Dosage'];
+            } else {
+              dosageList =
+                  getDosage(weight.toDouble(), mainDrug['Dosage'], isLbs);
+
+              dosageDisplay = dosageList[0];
+            }
 
             // concentration
             String concentrationDisplay = "";
-
             if (mainDrug['Concentration'] == null) {
               concentrationDisplay = "N/A";
             } else if (mainDrug['Concentration'] is String) {
               concentrationDisplay = mainDrug['Concentration'];
             } else {
-              concentrationDisplay = getConcentration(mainDrug['Concentration'],
-                  dosageList[2], dosageList[3], dosageList[1]);
+              mainDrug['Seperated'] == true
+                  ? concentrationDisplay = getConcentrationSeperated(
+                      mainDrug['Concentration'], dosageList[1])
+                  : concentrationDisplay = getConcentration(
+                      mainDrug['Concentration'], dosageList[1]);
             }
 
             // notes
@@ -242,8 +372,7 @@ class DosagePage extends StatelessWidget {
                     child: Column(
                       mainAxisAlignment: MainAxisAlignment.center,
                       children: [
-                        // Text("Animal: $animal",
-                        //     style: const TextStyle(fontSize: 20)),
+                        //animal
                         Wrap(alignment: WrapAlignment.center, children: [
                           const Text("Animal: ",
                               textAlign: TextAlign.center,
@@ -253,8 +382,8 @@ class DosagePage extends StatelessWidget {
                               style: const TextStyle(
                                   fontSize: 20, fontWeight: FontWeight.bold))
                         ]),
-                        // Text("Drug: $drugname",
-                        //     style: const TextStyle(fontSize: 20)),
+
+                        //drug
                         Wrap(alignment: WrapAlignment.center, children: [
                           const Text("Drug: ",
                               textAlign: TextAlign.center,
@@ -265,10 +394,10 @@ class DosagePage extends StatelessWidget {
                                   fontSize: 20, fontWeight: FontWeight.bold))
                         ]),
                         weightText,
+
                         const SizedBox(height: 50),
-                        // Text("Recommended Dosage: $dosageDisplay",
-                        //     textAlign: TextAlign.center,
-                        //     style: const TextStyle(fontSize: 20)),
+
+                        //dosage
                         Wrap(alignment: WrapAlignment.center, children: [
                           const Text("Recommended Dosage: ",
                               textAlign: TextAlign.center,
@@ -279,11 +408,10 @@ class DosagePage extends StatelessWidget {
                                   fontSize: 20, fontWeight: FontWeight.bold))
                         ]),
                         const SizedBox(height: 20),
-                        // Text("Concentration: $concentrationDisplay",
-                        //     textAlign: TextAlign.center,
-                        //     style: const TextStyle(fontSize: 20)),
+
+                        //concentration
                         Wrap(alignment: WrapAlignment.center, children: [
-                          const Text("Concentration: ",
+                          const Text("Concentration:",
                               textAlign: TextAlign.center,
                               style: TextStyle(fontSize: 20)),
                           Text(concentrationDisplay,
@@ -291,7 +419,10 @@ class DosagePage extends StatelessWidget {
                               style: const TextStyle(
                                   fontSize: 20, fontWeight: FontWeight.bold))
                         ]),
+
                         const SizedBox(height: 25),
+
+                        //notes
                         Padding(
                           padding: const EdgeInsets.all(
                               16.0), // add desired padding value here
@@ -305,9 +436,6 @@ class DosagePage extends StatelessWidget {
                                 style: const TextStyle(
                                     fontSize: 20, fontWeight: FontWeight.bold))
                           ]),
-                          //Text("Notes: $notesDisplay",
-                          //     textAlign: TextAlign.center,
-                          //     style: const TextStyle(fontSize: 20)),
                         ),
                       ],
                     ),
